refact: 페이지 별 채팅 입력창 컴포넌트 분리 및 공통 textarea 로직 추출
